refactoring page.tsx to make the file size smaller / break it up into mo
