+++ src/app/page.tsx
@@ -1,44 +1,39 @@
 
-
 'use client';
 
 import { useState, useEffect, useRef, useCallback } from 'react';
 import Image from 'next/image';
 import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
-import { Progress } from "@/components/ui/progress";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { useToast } from "@/hooks/use-toast";
-import { Wallet, Bed, Mail, Zap, Loader, KeyRound, Sprout, Network, ShoppingCart, BrainCircuit, HardDrive, FileUp, AlertTriangle, Copy, ShieldCheck, UploadCloud, Camera, Upload, TestTube, FilePlus2, CheckCircle2, UserCog, FileText, Activity, Clock, BarChart2, Lock, PiggyBank, Image as ImageIcon, ExternalLink, GraduationCap } from 'lucide-react';
-import DewDropIcon from '@/components/icons/DewDropIcon';
-import FlowerIcon from '@/components/icons/FlowerIcon';
+import { useToast } from '@/hooks/use-toast';
+import { Wallet, Bed, Loader, KeyRound, Sprout, UploadCloud, Camera, Upload, TestTube, FilePlus2, CheckCircle2, FileText, Image as ImageIcon, ExternalLink } from 'lucide-react';
 import { cn } from '@/lib/utils';
-import { Checkbox } from '@/components/ui/checkbox';
 import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
 import { detectSleepingSurface } from '@/ai/flows/detect-sleeping-surface-flow';
 import { scoreDataContribution } from '@/ai/flows/score-data-contribution-flow';
-import { ChartContainer, ChartTooltip, ChartTooltipContent, ChartConfig } from "@/components/ui/chart"
-import { BarChart, Bar, LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip as RechartsTooltip, Legend, ResponsiveContainer } from 'recharts';
 import { useWalletSelector } from '@/components/WalletProvider';
 import { CONTRACT_ID } from '@/lib/constants';
 import { utils, providers } from 'near-api-js';
 import type { CodeResult, FinalExecutionOutcome } from "near-api-js/lib/providers/provider";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
-import { Tooltip, TooltipProvider, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
-import FlowerIconTwo from '@/components/icons/FlowerIconTwo';
-import FlowerIconThree from '@/components/icons/FlowerIconThree';
+import { TooltipProvider } from '@/components/ui/tooltip';
 import TutorialDialog from '@/components/TutorialDialog';
+import Header from '@/components/app/Header';
+import AdminDashboard from '@/components/app/AdminDashboard';
+import ProofOfRest from '@/components/app/ProofOfRest';
+import ProofOfAction from '@/components/app/ProofOfAction';
+import ActionGarden from '@/components/app/ActionGarden';
+import DeviceStore from '@/components/app/DeviceStore';
+import DataContribution from '@/components/app/DataContribution';
+import SwarmStorage from '@/components/app/SwarmStorage';
+import SleepLog from '@/components/app/SleepLog';
+import LiveMotion from '@/components/app/LiveMotion';
 
 
-type MotionDataPoint = {
+export type MotionDataPoint = {
   time: number;
   magnitude: number;
 };
 
-type JournalEntry = {
+export type JournalEntry = {
   id: number;
   date: string;
   sleep: string;
@@ -46,7 +41,7 @@ type JournalEntry = {
   motionData?: MotionDataPoint[];
 };
 
-type PairedDataEntry = {
+export type PairedDataEntry = {
   id: number;
   date: string;
   fnirsFile: string;
@@ -54,60 +49,41 @@ type PairedDataEntry = {
   contributionScore: number;
 };
 
-type SwarmKeys = {
+export type SwarmKeys = {
   ethereumAddress: string;
   publicKey: string;
   password: string;
 } | null;
 
-type WhoopData = {
+export type WhoopData = {
     date: string;
     'Time in Bed (hours)': number;
     'Sleep (hours)': number;
 }[];
 
-type GardenFlower = {
+export type GardenFlower = {
     id: number;
     Icon: React.ElementType;
     unlocked: boolean;
 };
 
-const whoopChartConfig = {
-  timeInBed: {
-    label: "Time in Bed",
-    color: "hsl(var(--chart-1))",
-  },
-  sleep: {
-    label: "Sleep",
-    color: "hsl(var(--chart-2))",
-  },
-} satisfies ChartConfig
-
-const motionChartConfig = {
-    magnitude: {
-      label: "Movement",
-      color: "hsl(var(--primary))",
-    },
-} satisfies ChartConfig
-
-
-type StakerInfo = {
+export type StakerInfo = {
     amount: string; // Comes as a string from the contract
     bonus_approved: boolean;
 };
 
-type FileInfo = {
+export type FileInfo = {
     name: string;
     size: string;
     rows: number;
     cols: number;
 } | null;
 
-type Campaign = 'chat_control' | 'sugar_tax' | 'sleep_compensation';
+export type Campaign = 'chat_control' | 'sugar_tax' | 'sleep_compensation';
 
-type CampaignState = 'idle' | 'email_pending' | 'verified';
+export type CampaignState = 'idle' | 'email_pending' | 'verified';
 
-const CAMPAIGN_DETAILS: Record<Campaign, { title: string; description: string; subject: string }> = {
+export const CAMPAIGN_DETAILS: Record<Campaign, { title: string; description: string; subject: string }> = {
     chat_control: {
         title: "Stop Chat Control",
         description: "The EU is working on a law that would monitor all citizens' communications. Voice your opposition to this mass surveillance proposal.",
@@ -126,18 +102,18 @@ const CAMPAIGN_DETAILS: Record<Campaign, { title: string; description: string; s
 };
 
 
-const THIRTY_TGAS = "30000000000000";
-type MotionStatus = 'still' | 'slight' | 'heavy';
+export const THIRTY_TGAS = "30000000000000";
+export type MotionStatus = 'still' | 'slight' | 'heavy';
 
 export default function Home() {
   const [intentionPoints, setIntentionPoints] = useState(250); // Start with enough points to test
   const [journalEntries, setJournalEntries] = useState<JournalEntry[]>([]);
   const [gardenFlowers, setGardenFlowers] = useState<GardenFlower[]>([
-    { id: 1, Icon: FlowerIcon, unlocked: false },
-    { id: 2, Icon: FlowerIconTwo, unlocked: false },
-    { id: 3, Icon: FlowerIconThree, unlocked: false },
-    { id: 4, Icon: FlowerIcon, unlocked: false },
-    { id: 5, Icon: FlowerIconTwo, unlocked: false },
+    { id: 1, Icon: require('@/components/icons/FlowerIcon').default, unlocked: false },
+    { id: 2, Icon: require('@/components/icons/FlowerIconTwo').default, unlocked: false },
+    { id: 3, Icon: require('@/components/icons/FlowerIconThree').default, unlocked: false },
+    { id: 4, Icon: require('@/components/icons/FlowerIcon').default, unlocked: false },
+    { id: 5, Icon: require('@/components/icons/FlowerIconTwo').default, unlocked: false },
   ]);
   const [hasFnirsDevice, setHasFnirsDevice] = useState(false);
   const [hasAbbottDevice, setHasAbbottDevice] = useState(false);
@@ -1263,394 +1239,25 @@ export default function Home() {
     navigator.clipboard.writeText(text);
     toast({ title: 'Copied to clipboard!' });
   }
-  
-  const getStateDescription = (state: typeof appState) => {
-    switch (state) {
-      case 'analyzing_photo':
-        return { icon: <BrainCircuit className="animate-pulse text-primary" />, text: 'Analyzing photo for sleeping surface...' };
-      case 'sleeping':
-        return { icon: <Bed className="animate-pulse text-primary" />, text: 'Capturing sleep sensor data...' };
-      case 'generating_sleep_proof':
-        return { icon: <KeyRound className="animate-spin text-primary" />, text: 'Generating ZK-Proof of Rest...' };
-      case 'minting_dew':
-        return { icon: <Zap className="futuristic-glow text-primary" />, text: 'Withdrawing your commitment...' };
-      case 'generating_action_proof':
-        return { icon: <KeyRound className="animate-spin text-primary" />, text: 'Generating ZK-Proof of Action...' };
-      case 'planting_seed':
-        return { icon: <Sprout className="sprout text-primary" />, text: 'Verifying on Civic Action Registry...' };
-      case 'uploading_data':
-        return { icon: <UploadCloud className="animate-pulse text-primary" />, text: 'Analyzing and scoring your contribution...' };
-      default:
-        return { icon: null, text: '' };
-    }
-  };
-
-  const ProgressDisplay = ({ state, inCard = true } : { state: typeof appState; inCard?: boolean }) => {
-    const statesToShow = [
-      'analyzing_photo', 'sleeping', 'generating_sleep_proof', 'minting_dew',
-      'generating_action_proof', 'planting_seed', 'uploading_data'
-    ];
-    if (!statesToShow.includes(state)) {
-      return null;
-    }
-
-    const { icon, text } = getStateDescription(state);
-
-    const content = (
-        <div className="space-y-3 fade-in">
-            <div className="flex items-center gap-3 text-sm font-medium">
-                {icon}
-                <span>{text}</span>
-            </div>
-            <Progress value={progress} className="w-full h-2" />
-        </div>
-    );
-    
-    if (inCard) {
-      return <div className="mt-4 p-4 bg-secondary/50 rounded-lg">{content}</div>;
-    }
-    
-    return content
-  };
-
 
   if (isLoading) {
     return <div className="flex h-screen w-full items-center justify-center bg-background"><Loader className="h-12 w-12 animate-spin text-primary" /></div>;
   }
   
-  const renderAdminDashboard = () => (
-    <Card className="lg:col-span-2 slide-in-from-bottom transition-all hover:shadow-primary/5">
-        <CardHeader>
-            <CardTitle className="font-headline text-2xl flex items-center gap-3"><UserCog className="text-primary"/>Admin Dashboard</CardTitle>
-            <CardDescription>Approve staking bonuses for users and monitor the reward pool.</CardDescription>
-        </CardHeader>
-        <CardContent className="space-y-6">
-             <div className="p-4 rounded-lg bg-secondary flex flex-col sm:flex-row items-start sm:items-center justify-between gap-4">
-                <div className="flex items-center gap-3">
-                    <PiggyBank className="text-primary" size={24}/>
-                    <div >
-                        <p className="font-semibold">Reward Pool Balance</p>
-                        <p className="text-sm text-muted-foreground">Funds available for bonuses.</p>
-                    </div>
-                </div>
-                {rewardPoolBalance !== null ? (
-                    <p className="font-bold text-xl text-primary">{utils.format.formatNearAmount(rewardPoolBalance, 4)} NEAR</p>
-                ) : (
-                    <Loader className="animate-spin" size={20} />
-                )}
-             </div>
-             <div className="space-y-4">
-                <Label htmlFor="staker-id">Approve Bonus for Staker</Label>
-                <div className="flex items-start gap-2">
-                    <div className="flex-grow space-y-2">
-                        <Input 
-                            id="staker-id" 
-                            value={stakerIdToApprove}
-                            onChange={(e) => setStakerIdToApprove(e.target.value)}
-                            placeholder="e.g. user.near"
-                        />
-                        {isCheckingAddress && <p className="text-sm text-muted-foreground flex items-center gap-2"><Loader className="animate-spin" size={16}/> Checking...</p>}
-                        {infoForAddress && (
-                            <Alert>
-                                <AlertTitle className="flex items-center gap-2">
-                                    {infoForAddress.bonus_approved ? <CheckCircle2 className="text-green-500" /> : <AlertTriangle className="text-yellow-500" />}
-                                    Staker Information
-                                </AlertTitle>
-                                <AlertDescription>
-                                    <p>Stake: {utils.format.formatNearAmount(infoForAddress.amount, 4)} NEAR</p>
-                                    <p>Bonus Approved: {infoForAddress.bonus_approved ? 'Yes' : 'No'}</p>
-                                </AlertDescription>
-                            </Alert>
-                        )}
-                        {!isCheckingAddress && stakerIdToApprove.length > 5 && !infoForAddress && (
-                             <Alert variant="destructive">
-                                <AlertTriangle className="h-4 w-4" />
-                                <AlertTitle>Staker Not Found</AlertTitle>
-                                <AlertDescription>
-                                    This account has no stake in this contract.
-                                </AlertDescription>
-                            </Alert>
-                        )}
-                    </div>
-                    <Button 
-                        onClick={() => handleApproveBonus(stakerIdToApprove)} 
-                        disabled={!infoForAddress || infoForAddress.bonus_approved}
-                    >
-                        Approve Bonus
-                    </Button>
-                </div>
-             </div>
-              <div className="space-y-4 pt-4 border-t">
-                <Label htmlFor="deposit-amount">Deposit to Reward Pool</Label>
-                 <div className="flex items-stretch gap-2">
-                    <div className="flex-grow">
-                        <Label htmlFor="deposit-amount" className="sr-only">Deposit Amount (NEAR)</Label>
-                        <Input 
-                            id="deposit-amount" 
-                            type="number" 
-                            value={depositAmount}
-                            onChange={(e) => setDepositAmount(e.target.value)}
-                            placeholder="Amount in NEAR"
-                            className="[appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none"
-                        />
-                    </div>
-                    <Button onClick={handleDepositRewardFunds} disabled={!depositAmount || Number(depositAmount) <= 0}>
-                        Deposit NEAR
-                    </Button>
-                </div>
-             </div>
-        </CardContent>
-    </Card>
-  );
-
-  const FileInfoDisplay = ({ info }: { info: FileInfo }) => {
-    if (!info) return null;
-    return (
-        <Alert variant="default" className="mt-2 text-xs">
-            <FileText className="h-4 w-4" />
-            <AlertTitle className="font-semibold">{info.name}</AlertTitle>
-            <AlertDescription className="flex justify-between items-center">
-                <span>{info.size}</span>
-                <span>{info.rows} rows</span>
-                <span>{info.cols} columns</span>
-            </AlertDescription>
-        </Alert>
-    )
-  }
-
-  const MotionChart = ({ data, title = "Motion Analysis" }: { data: MotionDataPoint[]; title?: string; }) => {
-    if (!data || data.length === 0) {
-      return <p className="text-sm text-muted-foreground text-center py-4">No motion data recorded for this session.</p>;
-    }
-  
-    return (
-      <div className="mt-4">
-        <h4 className="font-headline text-lg mb-2">{title}</h4>
-        <ChartContainer config={motionChartConfig} className="h-[150px] w-full">
-          <LineChart data={data} margin={{ top: 5, right: 20, bottom: 5, left: 0 }}>
-            <CartesianGrid vertical={false} strokeDasharray="3 3" />
-            <XAxis 
-                dataKey="time" 
-                type="number" 
-                domain={['dataMin', 'dataMax']} 
-                tickFormatter={(tick) => `${Math.round(tick)}s`}
-                tickLine={false}
-                axisLine={false}
-                tickMargin={8}
-                />
-            <YAxis
-                tickLine={false}
-                axisLine={false}
-                tickMargin={8}
-                tickFormatter={(tick) => tick.toFixed(1)}
-                label={{ value: 'm/s²', angle: -90, position: 'insideLeft', offset: 10, style: { textAnchor: 'middle' } }}
-            />
-            <RechartsTooltip 
-                cursor={{ strokeDasharray: '3 3' }} 
-                content={({ active, payload, label }) => 
-                    active && payload && payload.length ? (
-                    <div className="rounded-lg border bg-background p-2 shadow-sm">
-                      <div className="grid grid-cols-2 gap-2">
-                        <div className="flex flex-col">
-                          <span className="text-[0.70rem] uppercase text-muted-foreground">Time</span>
-                          <span className="font-bold text-muted-foreground">{label.toFixed(1)}s</span>
-                        </div>
-                        <div className="flex flex-col">
-                          <span className="text-[0.70rem] uppercase text-muted-foreground">Movement</span>
-                          <span className="font-bold text-primary">{payload[0].value?.toFixed(2)}</span>
-                        </div>
-                      </div>
-                    </div>
-                  ) : null
-                }
-            />
-            <Line type="monotone" dataKey="magnitude" stroke="var(--color-magnitude)" strokeWidth={2} dot={false} name="Movement"/>
-          </LineChart>
-        </ChartContainer>
-      </div>
-    );
-  };
-  
-  const LiveMotionStatus = () => {
-    const statusConfig = {
-      still: {
-        color: "bg-green-500",
-        text: "Still",
-      },
-      slight: {
-        color: "bg-yellow-500",
-        text: "Slight Movement",
-      },
-      heavy: {
-        color: "bg-red-500",
-        text: "Heavy Movement",
-      },
-    };
-
-    const currentStatus = statusConfig[liveMotionStatus];
-
-    return (
-      <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5" style={{ animationDelay: '300ms' }}>
-        <CardHeader>
-          <CardTitle className="font-headline text-2xl flex items-center gap-2"><BarChart2 className="text-primary" />Live Sensor Data</CardTitle>
-          <CardDescription>Movement data from your device's accelerometer.</CardDescription>
-        </CardHeader>
-        <CardContent>
-          {hasMotionSensor === false && <p className="text-sm text-muted-foreground text-center py-4">Motion sensors not available or permission denied on this device.</p>}
-          {hasMotionSensor === null && <div className="flex items-center justify-center p-8 space-x-4"><Loader className="h-8 w-8 animate-spin text-primary" /><p>Accessing sensors...</p></div>}
-          {hasMotionSensor === true && (
-            <div className="flex items-center justify-center p-8 space-x-4">
-              <div className={cn("h-6 w-6 rounded-full transition-colors", currentStatus.color)} />
-              <p className="font-medium text-lg">{currentStatus.text}</p>
-            </div>
-          )}
-        </CardContent>
-      </Card>
-    );
-  };
-
-  const renderSleepLog = () => (
-     <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '300ms'}}>
-        <CardHeader>
-            <div className='flex items-start justify-between gap-4'>
-                <div>
-                    <CardTitle className="font-headline text-2xl">Sleep Log</CardTitle>
-                    <CardDescription>Your private, encrypted records of rest.</CardDescription>
-                </div>
-                <Button variant="outline" size="sm" onClick={() => whoopInputRef.current?.click()}>
-                    <UploadCloud className="mr-2 h-4 w-4" />
-                    Import
-                </Button>
-            </div>
-        </CardHeader>
-        <CardContent>
-            {whoopData.length > 0 && (
-                <div className="mb-6">
-                    <h4 className="text-lg font-headline mb-2">Recent Sleep Performance</h4>
-                    <ChartContainer config={whoopChartConfig} className="h-[200px] w-full">
-                        <BarChart accessibilityLayer data={whoopData} margin={{ top: 20, right: 20, bottom: 5, left: 0 }}>
-                            <CartesianGrid vertical={false} />
-                            <XAxis dataKey="date" tickLine={false} tickMargin={10} axisLine={false} />
-                            <YAxis unit="h" tickLine={false} axisLine={false} />
-                            <RechartsTooltip cursor={false} content={<ChartTooltipContent />} />
-                            <Legend />
-                            <Bar dataKey="Time in Bed (hours)" name="Time in Bed" fill="var(--color-timeInBed)" radius={4} />
-                            <Bar dataKey="Sleep (hours)" name="Sleep" fill="var(--color-sleep)" radius={4} />
-                        </BarChart>
-                    </ChartContainer>
-                </div>
-            )}
-            <ScrollArea className="h-[280px] pr-4">
-                <div className="space-y-4">
-                    {journalEntries.map(entry => (
-                        <div key={entry.id} className="flex flex-col gap-4 rounded-lg border p-3 bg-card hover:bg-secondary/50 transition-colors" data-ai-hint="bed bedroom">
-                            <div className="flex items-center gap-4">
-                                <Image src={entry.imageUrl} alt="A photo of a bed" width={80} height={60} className="rounded-md object-cover aspect-[4/3]" data-ai-hint="night sleep" />
-                                <div className="flex-grow">
-                                    <p className="font-semibold">{entry.date}</p>
-                                    <p className="text-sm text-primary">{entry.sleep}</p>
-                                </div>
-                            </div>
-                            {entry.motionData && entry.motionData.length > 0 && <MotionChart data={entry.motionData} />}
-                        </div>
-                    ))}
-                    {journalEntries.length === 0 && <p className="text-center text-muted-foreground pt-16">Complete a sleep verification to start your log.</p>}
-                </div>
-            </ScrollArea>
-        </CardContent>
-    </Card>
-  );
-
-  const renderLockedSleepLog = () => (
-     <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5 relative overflow-hidden" style={{animationDelay: '300ms'}}>
-        <div className="absolute inset-0 bg-background/70 backdrop-blur-sm z-10 flex flex-col items-center justify-center p-8 gap-4">
-             <div className="flex items-center gap-2 text-lg font-headline">
-                <Lock className="text-primary" />
-                <span>Sleep Log Locked</span>
-             </div>
-             <p className="text-center text-muted-foreground">Unlock this feature to track your sleep history and import data from other services.</p>
-             <Button onClick={handleUnlockSleepLog} disabled={intentionPoints < 500} className="mt-2">
-                 <DewDropIcon className="mr-2" />
-                 Unlock for 500 Points
-             </Button>
-        </div>
-        <CardHeader>
-            <div className='flex items-start justify-between gap-4'>
-                <div>
-                    <CardTitle className="font-headline text-2xl">Sleep Log</CardTitle>
-                    <CardDescription>Your private, encrypted records of rest.</CardDescription>
-                </div>
-                <Button variant="outline" size="sm" disabled>
-                    <UploadCloud className="mr-2 h-4 w-4" />
-                    Import
-                </Button>
-            </div>
-        </CardHeader>
-        <CardContent>
-             <ScrollArea className="h-[280px] pr-4">
-                <div className="space-y-4">
-                    <p className="text-center text-muted-foreground pt-16">Unlock to see your sleep history.</p>
-                </div>
-             </ScrollArea>
-        </CardContent>
-    </Card>
-  )
-
-
   return (
     <TooltipProvider>
     <div className="min-h-screen w-full bg-background text-foreground fade-in">
       <TutorialDialog open={showTutorial} onOpenChange={setShowTutorial} />
-      <header className="sticky top-0 z-10 border-b border-border/50 bg-background/80 backdrop-blur-sm">
-        <div className="container mx-auto flex h-16 items-center justify-between px-4 sm:px-6 lg:px-8">
-            <a href="/" onClick={(e) => {
-                if(!window.confirm("Are you sure? This will refresh the page and you may lose your current progress.")) {
-                    e.preventDefault();
-                }
-            }}>
-                <h1 className="font-headline text-xl md:text-2xl text-primary">think2earn</h1>
-            </a>
-          <div className="flex items-center gap-2 md:gap-4">
-            <div className="flex items-center gap-2 rounded-full border border-border/50 bg-card px-3 md:px-4 py-2 text-sm shadow-sm">
-                <DewDropIcon className="h-5 w-5 text-primary" />
-                <span className="text-base md:text-lg">{intentionPoints}</span>
-                <span className="text-muted-foreground hidden sm:inline">Intention Points</span>
-            </div>
-            
-            <Tooltip>
-                <TooltipTrigger asChild>
-                    <Button variant="ghost" size="icon" onClick={() => setShowTutorial(true)}>
-                        <GraduationCap className="h-5 w-5 text-muted-foreground hover:text-primary" />
-                        <span className="sr-only">Show Tutorial</span>
-                    </Button>
-                </TooltipTrigger>
-                <TooltipContent>
-                    <p>Show Tutorial</p>
-                </TooltipContent>
-            </Tooltip>
-
-            {walletConnected ? (
-                <div className="flex items-center gap-2">
-                    <div className="flex items-center gap-2 rounded-full border border-border/50 bg-card pl-3 pr-2 md:pl-4 py-1 text-sm shadow-sm">
-                        <div className='text-right'>
-                           <p className="font-mono text-xs md:text-sm truncate max-w-[100px] sm:max-w-none">{signedAccountId}</p>
-                           <p className='text-xs text-primary font-semibold'>{accountBalance !== null ? `${accountBalance} NEAR` : <Loader size={12} className="inline-block animate-spin"/>}</p>
-                        </div>
-                        <Button variant="ghost" size="icon" className="h-8 w-8" onClick={logOut}>
-                            <Wallet className="h-5 w-5 text-muted-foreground hover:text-primary" />
-                            <span className='sr-only'>Logout</span>
-                        </Button>
-                    </div>
-                </div>
-            ) : (
-                <Button onClick={logIn} disabled={isLoggingIn}>
-                    <Wallet className="mr-2 h-4 w-4" />
-                    {isLoggingIn ? 'Connecting...' : 'Connect NEAR Wallet'}
-                </Button>
-            )}
-          </div>
-        </div>
-      </header>
+      <Header
+        intentionPoints={intentionPoints}
+        walletConnected={walletConnected}
+        signedAccountId={signedAccountId}
+        accountBalance={accountBalance}
+        isLoggingIn={isLoggingIn}
+        logIn={logIn}
+        logOut={logOut}
+        setShowTutorial={setShowTutorial}
+      />
       
       <main className="container mx-auto p-4">
       <Input type="file" accept=".csv" ref={whoopInputRef} onChange={handleWhoopImport} className="hidden" />
@@ -1665,427 +1272,134 @@ export default function Home() {
             <div className="lg:col-span-2 space-y-6">
 
                 {appState === 'taking_photo' && (
-                    <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5">
-                        <CardHeader>
-                            <CardTitle className="font-headline text-2xl">Provide Evidence of Rest</CardTitle>
-                            <CardDescription>Take a photo of your bed or upload one from your gallery.</CardDescription>
-                        </CardHeader>
-                        <CardContent className="space-y-4">
-                            <div className="w-full aspect-video bg-secondary rounded-md overflow-hidden flex items-center justify-center relative">
-                                {uploadedImage ? (
-                                    <Image src={uploadedImage.url} alt="Uploaded bed photo" layout="fill" objectFit="cover" />
-                                ) : (
-                                    <>
-                                        <video ref={videoRef} className={cn("h-full w-full object-cover", { 'hidden': hasCameraPermission !== true })} autoPlay muted playsInline />
-                                        {hasCameraPermission === false && 
-                                            <Alert variant="default" className='m-4'>
-                                                <AlertTriangle className="h-4 w-4" />
-                                                <AlertTitle>Camera Unavailable</AlertTitle>
-                                                <AlertDescription>
-                                                    No problem. Please upload a photo from your gallery or use the default image.
-                                                </AlertDescription>
-                                            </Alert>
-                                        }
-                                        {hasCameraPermission === null && !videoRef.current?.srcObject && <Loader className="h-8 w-8 animate-spin text-primary" />}
-                                    </>
-                                )}
-                            </div>
-
-                            {uploadedImage && (
-                                <Alert variant="default">
-                                    <AlertTriangle className="h-4 w-4" />
-                                    <AlertTitle>Image Selected</AlertTitle>
-                                    <AlertDescription>
-                                        {uploadedImage.date}
-                                    </AlertDescription>
-                                </Alert>
-                            )}
-                            <canvas ref={canvasRef} className="hidden" />
-                            <Input type="file" accept="image/*" ref={fileInputRef} onChange={handleFileUpload} className="hidden" />
-                            {isVerifyingSleep && <ProgressDisplay state={appState} inCard={false} />}
-                        </CardContent>
-                        <CardFooter className="flex flex-col sm:flex-row gap-2">
-                           <Button variant="outline" onClick={handleUseDefaultPhoto} className="w-full">
-                                <ImageIcon className="mr-2 h-4 w-4"/>
-                                Use Default Photo
-                            </Button>
-                             <Button variant="outline" onClick={() => takePhoto()} disabled={!hasCameraPermission} className="w-full">
-                                <Camera className="mr-2 h-4 w-4"/>
-                                Take Photo
-                            </Button>
-                            <Button variant="outline" onClick={() => fileInputRef.current?.click()} className="w-full">
-                                <Upload className="mr-2 h-4 w-4"/>
-                                Upload from Gallery
-                            </Button>
-                            <div className='flex-grow'/>
-                            <Button onClick={handleConfirmPhoto} disabled={!uploadedImage || isVerifyingSleep} className="w-full sm:w-auto">
-                                <ShieldCheck className="mr-2 h-4 w-4"/>
-                                Confirm and Verify
-                            </Button>
-                            <Button variant="ghost" onClick={() => { setAppState('idle'); setUploadedImage(null); }} className="w-full sm:w-auto">Cancel</Button>
-                        </CardFooter>
-                    </Card>
+                    <ProofOfRest
+                      appState={appState}
+                      progress={progress}
+                      uploadedImage={uploadedImage}
+                      videoRef={videoRef}
+                      canvasRef={canvasRef}
+                      fileInputRef={fileInputRef}
+                      hasCameraPermission={hasCameraPermission}
+                      isVerifyingSleep={isVerifyingSleep}
+                      handleUseDefaultPhoto={handleUseDefaultPhoto}
+                      takePhoto={takePhoto}
+                      handleFileUpload={handleFileUpload}
+                      handleConfirmPhoto={handleConfirmPhoto}
+                      setAppState={setAppState}
+                      setUploadedImage={setUploadedImage}
+                    />
                 )}
 
                 {appState !== 'taking_photo' && (
                   <>
-                    {isAdmin && walletConnected && renderAdminDashboard()}
+                    {isAdmin && walletConnected && (
+                      <AdminDashboard
+                        rewardPoolBalance={rewardPoolBalance}
+                        stakerIdToApprove={stakerIdToApprove}
+                        setStakerIdToApprove={setStakerIdToApprove}
+                        isCheckingAddress={isCheckingAddress}
+                        infoForAddress={infoForAddress}
+                        handleApproveBonus={handleApproveBonus}
+                        depositAmount={depositAmount}
+                        setDepositAmount={setDepositAmount}
+                        handleDepositRewardFunds={handleDepositRewardFunds}
+                      />
+                    )}
 
-                    <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5">
-                      <CardHeader>
-                          <CardTitle className="font-headline text-2xl">Daily Positive Actions</CardTitle>
-                          <CardDescription>Generate proofs of your positive actions by making a commitment.</CardDescription>
-                      </CardHeader>
-                      <CardContent className="space-y-6">
-                          <div className="space-y-4 rounded-lg border p-4 hover:border-primary/20 transition-colors">
-                              <div className="flex items-center gap-3">
-                                  <Bed className="h-6 w-6 text-primary" />
-                                  <h3 className="font-headline text-lg">Proof of Rest</h3>
-                              </div>
-                              <p className="text-sm text-muted-foreground">Commit NEAR to verify your sleep. After verification, your commitment is returned with a bonus.</p>
-                              {rewardPoolBalance !== null && (
-                                  <div className="text-xs text-muted-foreground flex items-center gap-2">
-                                      <PiggyBank size={14}/>
-                                      <span>Available Rewards: {utils.format.formatNearAmount(rewardPoolBalance, 4)} NEAR</span>
-                                  </div>
-                              )}
-
-                              {stakerInfo && walletConnected ? (
-                                  <div className='p-4 bg-secondary rounded-md space-y-3'>
-                                      <div>
-                                          <p className='text-sm font-semibold'>You have <span className="font-bold text-primary">{utils.format.formatNearAmount(stakerInfo.amount, 4)} NEAR</span> committed.</p>
-                                          <p className="text-xs text-muted-foreground mt-1">Complete sleep verification to get it back with a bonus.</p>
-                                      </div>
-                                      <div className="flex items-center gap-2 text-sm">
-                                          <span>Bonus Status:</span>
-                                          {stakerInfo.bonus_approved ? (
-                                              <span className='font-medium text-green-600 flex items-center gap-1'><CheckCircle2 size={16}/> Approved</span>
-                                          ) : (
-                                              <Tooltip>
-                                                  <TooltipTrigger asChild>
-                                                      <span className='font-medium text-muted-foreground flex items-center gap-1 cursor-help'><Clock size={16} /> Pending</span>
-                                                  </TooltipTrigger>
-                                                  <TooltipContent>
-                                                      <p>Bonus is approved manually by the admin.<br/> It may take a few moments to be verified.</p>
-                                                  </TooltipContent>
-                                              </Tooltip>
-                                          )}
-                                      </div>
-                                      <div className="flex flex-col sm:flex-row gap-2">
-                                          <Button onClick={handleBeginSleepVerification} disabled={isVerifyingSleep || stakerInfo.bonus_approved} className="w-full">
-                                              Verify Sleep
-                                          </Button>
-                                          <Button onClick={handleWithdraw} disabled={isVerifyingSleep || !stakerInfo.bonus_approved} className="w-full" variant="outline">
-                                              Withdraw
-                                          </Button>
-                                      </div>
-                                      {isVerifyingSleep && <ProgressDisplay state={appState} />}
-                                  </div>
-                              ) : (
-                                  <div className="flex flex-col sm:flex-row items-stretch gap-2">
-                                      <div className="flex-grow">
-                                          <Label htmlFor="stake-amount" className="sr-only">Commitment (NEAR)</Label>
-                                          <Input id="stake-amount" type="number" value={stakeAmount} onChange={(e) => setStakeAmount(e.target.value)} placeholder="Commitment (NEAR)" className="[appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none" />
-                                      </div>
-                                      <Button onClick={handleStake} disabled={isVerifyingSleep || !walletConnected || Number(stakeAmount) <= 0} className="w-full sm:w-auto">
-                                          Commit & Begin
-                                      </Button>
-                                  </div>
-                              )}
-                              {isVerifyingSleep && !stakerInfo && <ProgressDisplay state={appState} />}
-                          </div>
-
-                          <div className="space-y-4 rounded-lg border p-4 hover:border-primary/20 transition-colors">
-                              <div className="flex items-center gap-3">
-                                  <Mail className="h-6 w-6 text-primary" />
-                                  <h3 className="font-headline text-lg">Proof of Action</h3>
-                              </div>
-                              <p className="text-sm text-muted-foreground">Select a campaign and send a signed email to prove you've taken action. Costs 10 Intention Points.</p>
-                              <RadioGroup value={selectedCampaign} onValueChange={(value) => setSelectedCampaign(value as Campaign)} className="space-y-2">
-                                  {Object.keys(CAMPAIGN_DETAILS).map((campaignKey) => {
-                                      const campaign = campaignKey as Campaign;
-                                      const details = CAMPAIGN_DETAILS[campaign];
-                                      const campaignState = campaignStates[campaign];
-
-                                      return (
-                                          <Label key={campaign} htmlFor={campaign} className={cn("flex flex-col gap-2 rounded-md border p-3 cursor-pointer", {'border-primary ring-2 ring-primary': selectedCampaign === campaign, 'opacity-50 cursor-not-allowed': campaignState === 'verified'})}>
-                                              <div className="flex items-start gap-3">
-                                                  <RadioGroupItem value={campaign} id={campaign} className="mt-1" disabled={campaignState === 'verified'} />
-                                                  <div className="flex-grow space-y-1">
-                                                      <p className="font-medium">{details.title}</p>
-                                                      <p className="text-sm text-muted-foreground">{details.description}</p>
-                                                  </div>
-                                              </div>
-                                              {selectedCampaign === campaign && campaignState !== 'verified' && (
-                                                  <div className="pt-2 pl-7">
-                                                      {campaignState === 'idle' && (
-                                                          <Button onClick={() => handleSendEmail(campaign)} disabled={isVerifyingAction || intentionPoints < 10} className="w-full" size="sm">
-                                                              <Mail className="mr-2"/>
-                                                              {intentionPoints < 10 ? 'Need 10 Points' : 'Send Email for 10 Points'}
-                                                          </Button>
-                                                      )}
-                                                      {campaignState === 'email_pending' && (
-                                                          <Button onClick={() => emailUploadRef.current?.click()} disabled={isVerifyingAction} className="w-full" size="sm" variant="outline">
-                                                              <Upload className="mr-2"/>
-                                                              Upload Signed Email (.eml/.txt)
-                                                          </Button>
-                                                      )}
-                                                  </div>
-                                              )}
-                                              {campaignState === 'verified' && (
-                                                  <div className="pt-2 pl-7">
-                                                      <Alert variant="default" className="border-green-500 text-green-700">
-                                                          <CheckCircle2 className="h-4 w-4 text-green-500" />
-                                                          <AlertTitle>Action Verified!</AlertTitle>
-                                                          <AlertDescription>
-                                                            You have successfully completed this action.
-                                                          </AlertDescription>
-                                                      </Alert>
-                                                  </div>
-                                              )}
-                                          </Label>
-                                      )
-                                  })}
-                              </RadioGroup>
-                              {isVerifyingAction && <ProgressDisplay state={appState} />}
-                          </div>
-                      </CardContent>
-                      </Card>
+                    <ProofOfRest
+                      appState={appState}
+                      progress={progress}
+                      rewardPoolBalance={rewardPoolBalance}
+                      stakerInfo={stakerInfo}
+                      walletConnected={walletConnected}
+                      isVerifyingSleep={isVerifyingSleep}
+                      stakeAmount={stakeAmount}
+                      setStakeAmount={setStakeAmount}
+                      handleStake={handleStake}
+                      handleBeginSleepVerification={handleBeginSleepVerification}
+                      handleWithdraw={handleWithdraw}
+                    />
+                    
+                    <ProofOfAction
+                        appState={appState}
+                        progress={progress}
+                        selectedCampaign={selectedCampaign}
+                        setSelectedCampaign={setSelectedCampaign}
+                        campaignStates={campaignStates}
+                        isVerifyingAction={isVerifyingAction}
+                        intentionPoints={intentionPoints}
+                        handleSendEmail={handleSendEmail}
+                        emailUploadRef={emailUploadRef}
+                    />
                   </>
                 )}
 
-                 <Card className="slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '100ms'}}>
-                    <CardHeader>
-                    <CardTitle className="font-headline text-2xl">My Action Garden</CardTitle>
-                    <CardDescription>A testament to your verified positive actions.</CardDescription>
-                    </CardHeader>
-                    <CardContent>
-                    <div className="grid grid-cols-5 sm:grid-cols-8 md:grid-cols-10 gap-4 p-4 rounded-lg min-h-[160px] bg-secondary/30 border border-border/50">
-                        {gardenFlowers.map((flower, i) => {
-                            const { Icon, unlocked } = flower;
-                            return (
-                                <Icon 
-                                    key={i} 
-                                    className={cn(
-                                        "h-10 w-10 transition-all duration-500",
-                                        unlocked ? "text-primary sprout" : "text-gray-300 opacity-50",
-                                    )} 
-                                    style={{ animationDelay: `${i * 50}ms` }}
-                                />
-                            )
-                        })}
-                        {gardenFlowers.every(f => f.unlocked) && <p className="col-span-full text-center text-muted-foreground self-center">Your garden is in full bloom! Keep up the great work.</p>}
-                    </div>
-                    </CardContent>
-                </Card>
-
-                 <Card className="lg:col-span-2 slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '400ms'}}>
-                    <CardHeader>
-                        <CardTitle className="font-headline text-2xl flex items-center gap-3"><ShoppingCart className="text-primary"/> Device Store</CardTitle>
-                        <CardDescription>Acquire the tools to contribute to glucose monitoring research.</CardDescription>
-                    </CardHeader>
-                    <CardContent className="grid sm:grid-cols-2 gap-4">
-                        <Card className="p-4 flex flex-col items-start gap-4 hover:bg-secondary/50 transition-colors">
-                            <Image src="https://placehold.co/600x400/000000/FFFFFF/png?text=fNIRS+Device" alt="Futuristic fNIRS armband wearable" width={600} height={400} className="rounded-lg self-center aspect-video object-cover" data-ai-hint="futuristic armband" />
-                            <div className="flex-grow space-y-2">
-                                <h3 className="font-headline text-lg">fNIRS Armband</h3>
-                                <p className="text-sm text-muted-foreground">Open-hardware fNIRS device for continuous, non-invasive data collection.</p>
-                                <div className="flex items-center gap-2 text-primary font-bold">
-                                    <DewDropIcon className="h-5 w-5"/>
-                                    <span>100 Intention Points</span>
-                                </div>
-                            </div>
-                            <Button onClick={() => handleAcquireDevice(100, 'fnirs')} disabled={intentionPoints < 100 || hasFnirsDevice} className="w-full">
-                            {hasFnirsDevice ? 'Acquired' : 'Acquire for 100 Points'}
-                            </Button>
-                        </Card>
-                        <Card className="p-4 flex flex-col items-start gap-4 hover:bg-secondary/50 transition-colors">
-                            <Image src="https://placehold.co/400x400/000000/FFFFFF/png?text=Glucose+Monitor" alt="Small, circular glucose monitoring device" width={400} height={400} className="rounded-lg self-center aspect-square object-cover" data-ai-hint="medical sensor" />
-                            <div className="flex-grow space-y-2">
-                                <h3 className="font-headline text-lg">Abbott Glucose Monitor</h3>
-                                <p className="text-sm text-muted-foreground">Certified medical device for providing baseline glucose data to train the model.</p>
-                                <div className="flex items-center gap-2 text-primary font-bold">
-                                <DewDropIcon className="h-5 w-5"/>
-                                <span>150 Intention Points</span>
-                                </div>
-                            </div>
-                            <Button onClick={() => handleAcquireDevice(150, 'abbott')} disabled={intentionPoints < 150 || hasAbbottDevice} className="w-full">
-                                {hasAbbottDevice ? 'Acquired' : 'Acquire for 150 Points'}
-                            </Button>
-                        </Card>
-                    </CardContent>
-                </Card>
+                <ActionGarden gardenFlowers={gardenFlowers} />
+                
+                <DeviceStore
+                    intentionPoints={intentionPoints}
+                    hasFnirsDevice={hasFnirsDevice}
+                    hasAbbottDevice={hasAbbottDevice}
+                    handleAcquireDevice={handleAcquireDevice}
+                />
 
                 {(hasFnirsDevice && hasAbbottDevice) && (
-                     <Card className="lg:col-span-2 slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '200ms'}}>
-                        <CardHeader>
-                            <CardTitle className="font-headline text-2xl flex items-center gap-3"><TestTube className="text-primary"/> Contribute Data</CardTitle>
-                            <CardDescription>Pair your fNIRS data with a certified glucose reading to help train the model.</CardDescription>
-                        </CardHeader>
-                        <form onSubmit={handleDataContribution}>
-                        <CardContent className="space-y-4">
-                            <div className="grid sm:grid-cols-2 gap-4">
-                                <div className="space-y-2">
-                                    <Label htmlFor="fnirs-file">1. Upload fNIRS Data</Label>
-                                    <Button type="button" variant="outline" className="w-full justify-start text-left font-normal" onClick={() => fnirsInputRef.current?.click()}>
-                                        <FilePlus2 className="mr-2" />
-                                        {fnirsFile ? <span className='truncate'>{fnirsFile.name}</span> : 'Select a .csv or .txt file'}
-                                    </Button>
-                                    <FileInfoDisplay info={fnirsInfo} />
-                                </div>
-                                <div className="space-y-2">
-                                    <Label htmlFor="glucose-level">2. Pair Glucose Reading File</Label>
-                                    <Button type="button" variant="outline" className="w-full justify-start text-left font-normal" onClick={() => glucoseInputRef.current?.click()}>
-                                        <FilePlus2 className="mr-2" />
-                                        {glucoseFile ? <span className="truncate">{glucoseFile.name}</span> : 'Select a .csv or .txt file'}
-                                    </Button>
-                                    <FileInfoDisplay info={glucoseInfo} />
-                                </div>
-                            </div>
-                             {isUploadingData && <ProgressDisplay state={appState} />}
-                        </CardContent>
-                        <CardFooter>
-                             <Button type="submit" className="w-full" disabled={!fnirsFile || !glucoseFile || isUploadingData}>
-                                {isUploadingData ? 'Submitting...' : 'Submit Paired Data'}
-                             </Button>
-                        </CardFooter>
-                        </form>
-                     </Card>
+                     <DataContribution
+                        appState={appState}
+                        progress={progress}
+                        fnirsFile={fnirsFile}
+                        glucoseFile={glucoseFile}
+                        fnirsInfo={fnirsInfo}
+                        glucoseInfo={glucoseInfo}
+                        fnirsInputRef={fnirsInputRef}
+                        glucoseInputRef={glucoseInputRef}
+                        handleFnirsUpload={handleFnirsUpload}
+                        handleGlucoseUpload={handleGlucoseUpload}
+                        handleDataContribution={handleDataContribution}
+                        isUploadingData={isUploadingData}
+                     />
                 )}
                 
                 {pairedDataHistory.length > 0 && (
-                     <Card className="lg-col-span-2 slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '300ms'}}>
-                        <CardHeader>
-                            <CardTitle className="font-headline text-2xl">Contribution History</CardTitle>
-                            <CardDescription>Your history of data contributions to the model.</CardDescription>
-                        </CardHeader>
-                        <CardContent>
-                            <Table>
-                                <TableHeader>
-                                    <TableRow>
-                                        <TableHead>Date</TableHead>
-                                        <TableHead>fNIRS File</TableHead>
-                                        <TableHead>Glucose (mg/dL)</TableHead>
-                                        <TableHead className="text-right">Contribution Score</TableHead>
-                                    </TableRow>
-                                </TableHeader>
-                                <TableBody>
-                                    {pairedDataHistory.map(entry => (
-                                        <TableRow key={entry.id}>
-                                            <TableCell>{entry.date}</TableCell>
-                                            <TableCell className="truncate max-w-[150px]">{entry.fnirsFile}</TableCell>
-                                            <TableCell>{entry.glucoseLevel}</TableCell>
-                                            <TableCell className="text-right font-medium text-primary">{entry.contributionScore}</TableCell>
-                                        </TableRow>
-                                    ))}
-                                </TableBody>
-                            </Table>
-                        </CardContent>
-                    </Card>
+                    <DataContribution.History pairedDataHistory={pairedDataHistory} />
                 )}
 
 
                 {showSwarmUI && (
-                    <Card className="lg:col-span-2 slide-in-from-bottom transition-all hover:shadow-primary/5" style={{animationDelay: '500ms'}}>
-                    <CardHeader>
-                        <CardTitle className="font-headline text-2xl flex items-center gap-3"><HardDrive className="text-primary"/> Sovereign Storage on Swarm</CardTitle>
-                        <CardDescription>Securely store your encrypted fNIRS data on a decentralized network.</CardDescription>
-                    </CardHeader>
-                    <CardContent className="space-y-4">
-                        {swarmState === 'idle' && (
-                        <>
-                            <p className="text-sm text-muted-foreground">Set up your personal, decentralized storage to own and control your health data. This involves creating a new Swarm account.</p>
-                            <Button onClick={handleSetupSwarm} className="w-full">
-                            Setup Swarm Storage
-                            </Button>
-                        </>
-                        )}
-
-                        {swarmState === 'generating_keys' && (
-                        <div className="flex items-center justify-center p-8 space-x-4">
-                            <Loader className="h-8 w-8 animate-spin text-primary" />
-                            <p>Generating your secure Swarm credentials...</p>
-                        </div>
-                        )}
-                        
-                        {swarmState === 'keys_generated' && swarmKeys && (
-                        <div className="space-y-4">
-                            <div className="p-4 rounded-lg bg-destructive/10 border border-destructive/50">
-                            <div className="flex items-start">
-                                <AlertTriangle className="h-5 w-5 mr-3 mt-1 text-destructive" />
-                                <div>
-                                <h4 className="font-bold text-destructive">Save These Credentials!</h4>
-                                <p className="text-sm text-destructive-foreground/90">This is generated locally in your browser so it cannot be retrieved by anyone else.</p>
-                                </div>
-                            </div>
-                            </div>
-                            <div className="space-y-2">
-                            <Label htmlFor="ethAddress">Gnosis Chain Address</Label>
-                            <div className="flex items-center gap-2">
-                                <Input id="ethAddress" value={swarmKeys.ethereumAddress} readOnly />
-                                <Button variant="ghost" size="icon" onClick={() => copyToClipboard(swarmKeys.ethereumAddress)}><Copy/></Button>
-                            </div>
-                            </div>
-                            <div className="space-y-2">
-                            <Label htmlFor="publicKey">Swarm Public Key</Label>
-                            <div className="flex items-center gap-2">
-                                <Input id="publicKey" value={swarmKeys.publicKey} readOnly />
-                                <Button variant="ghost" size="icon" onClick={() => copyToClipboard(swarmKeys.publicKey)}><Copy/></Button>
-                            </div>
-                            </div>
-                            <div className="space-y-2">
-                            <Label htmlFor="password">Password</Label>                            <div className="flex items-center gap-2">
-                                <Input id="password" value={swarmKeys.password} readOnly />
-                                <Button variant="ghost" size="icon" onClick={() => copyToClipboard(swarmKeys.password)}><Copy/></Button>
-                            </div>
-                            </div>
-                            <div className="flex items-center space-x-2 pt-2">
-                            <Checkbox id="terms" onCheckedChange={(checked) => setCredentialsSaved(checked as boolean)} />
-                            <label
-                                htmlFor="terms"
-                                className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-                            >
-                                I have saved my credentials in a secure place.
-                            </label>
-                            </div>
-                            <Button onClick={handleContinueFromKeys} disabled={!credentialsSaved} className="w-full">Continue</Button>
-                        </div>
-                        )}
-
-                        {swarmState === 'funding' && (
-                        <div className="space-y-4 rounded-lg border p-4">
-                            <h3 className="font-headline text-lg">1. Fund Your Account</h3>
-                            <p className="text-sm text-muted-foreground">To use Swarm, your Gnosis Chain address needs funds. Send ~0.01 xDAI (for gas fees) and ~0.2 xBZZ (for storage) to your address shown in the previous step.</p>
-                            <Button onClick={handleFunded} disabled={accountFunded} className="w-full">
-                                {accountFunded ? 'Account Funded' : 'I Have Funded My Account'}
-                            </Button>
-                        </div>
-                        )}
-                        {(swarmState === 'buying_stamps' || swarmState === 'ready_to_upload') && (
-                        <div className="space-y-4 rounded-lg border p-4">
-                            <h3 className="font-headline text-lg">2. Buy Postage Stamps</h3>
-                            <p className="text-sm text-muted-foreground">Postage stamps are required to upload data to Swarm. They cover the cost of storage for a specific duration.</p>
-                            <Button onClick={handleBuyStamps} disabled={stampsPurchased || swarmState !== 'buying_stamps'} className="w-full">
-                                {swarmState === 'ready_to_upload' ? 'Stamps Purchased!' : 'Purchase Postage Stamps'}
-                            </Button>
-                        </div>
-                        )}
-                        {swarmState === 'ready_to_upload' && (
-                        <div className="space-y-4 rounded-lg border p-4 bg-primary/10">
-                            <h3 className="font-headline text-lg flex items-center gap-3"><FileUp/> Ready to Upload</h3>
-                            <p className="text-sm text-muted-foreground">Your Swarm storage is set up. You can now upload your encrypted fNIRS data sessions.</p>
-                            <Button className="w-full">Upload fNIRS Data</Button>
-                        </div>
-                        )}
-                    </CardContent>
-                    </Card>
+                    <SwarmStorage
+                        swarmState={swarmState}
+                        setSwarmState={setSwarmState}
+                        swarmKeys={swarmKeys}
+                        setSwarmKeys={setSwarmKeys}
+                        credentialsSaved={credentialsSaved}
+                        setCredentialsSaved={setCredentialsSaved}
+                        accountFunded={accountFunded}
+                        setAccountFunded={setAccountFunded}
+                        stampsPurchased={stampsPurchased}
+                        setStampsPurchased={setStampsPurchased}
+                        handleSetupSwarm={handleSetupSwarm}
+                        handleContinueFromKeys={handleContinueFromKeys}
+                        handleFunded={handleFunded}
+                        handleBuyStamps={handleBuyStamps}
+                        copyToClipboard={copyToClipboard}
+                    />
                 )}
             </div>
             
             {/* Side Column */}
             <div className="lg:col-span-1 space-y-6">
-                {isSleepLogUnlocked ? renderSleepLog() : renderLockedSleepLog()}
-                <LiveMotionStatus />
+                <SleepLog
+                    isSleepLogUnlocked={isSleepLogUnlocked}
+                    intentionPoints={intentionPoints}
+                    whoopInputRef={whoopInputRef}
+                    whoopData={whoopData}
+                    journalEntries={journalEntries}
+                    handleUnlockSleepLog={handleUnlockSleepLog}
+                />
+                <LiveMotion
+                    liveMotionStatus={liveMotionStatus}
+                    hasMotionSensor={hasMotionSensor}
+                />
             </div>
         </div>
       </main>
@@ -2093,3 +1407,4 @@ export default function Home() {
     </TooltipProvider>
   );
 }
+
